+++ dotnet/samples/GettingStartedWithAgents/AzureAIAgent/Step01_AzureAIAgent.cs
@@ -14,7 +14,7 @@ namespace GettingStarted;
 /// This example demonstrates similarity between using <see cref="AzureAIAgent"/>
 /// and <see cref="ChatCompletionAgent"/> (see: Step 2).
 /// </summary>
-public class Step12_Azure(ITestOutputHelper output) : BaseAgentsTest(output)
+public class Step01_AzureAIAgent(ITestOutputHelper output) : BaseAgentsTest(output)
 {
     private const string HostName = "Host";
     private const string HostInstructions = "Answer questions about the menu.";
