+++ dotnet/samples/GettingStartedWithAgents/OpenAIAssistant/Step03_AssistantTool_CodeInterpreter.cs
@@ -8,7 +8,7 @@ namespace GettingStarted;
 /// <summary>
 /// Demonstrate using code-interpreter on <see cref="OpenAIAssistantAgent"/> .
 /// </summary>
-public class Step10_AssistantTool_CodeInterpreter(ITestOutputHelper output) : BaseAgentsTest(output)
+public class Step03_AssistantTool_CodeInterpreter(ITestOutputHelper output) : BaseAgentsTest(output)
 {
     [Fact]
     public async Task UseCodeInterpreterToolWithAssistantAgentAsync()
