+++ dotnet/samples/GettingStartedWithAgents/OpenAIAssistant/Step04_AssistantTool_FileSearch.cs
@@ -12,7 +12,7 @@ namespace GettingStarted;
 /// <summary>
 /// Demonstrate using code-interpreter on <see cref="OpenAIAssistantAgent"/> .
 /// </summary>
-public class Step11_AssistantTool_FileSearch(ITestOutputHelper output) : BaseAgentsTest(output)
+public class Step04_AssistantTool_FileSearch(ITestOutputHelper output) : BaseAgentsTest(output)
 {
     [Fact]
     public async Task UseFileSearchToolWithAssistantAgentAsync()
